+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# 0.3.0 - Pre-release (2019-08-24)
+
+- Can add additional informations to _species_: uses, ornemental, horticulture.
+- Page to show species informations.
+
 # 0.2.0 - Pre-release (2019-08-24)
 
 - Can add additional informations to _species_: architecture, functions.